Override Tools > Port menu label for mEDBG devices
Arduino megaAVR Boards associates the VID/PID of the mEDBG with the Arduino Uno WiFi Rev2 board, which causes all mEDBG boards to be labeled as "Arduino Uno WiFi Rev2" in the Arduino IDE's Tools > Port menu, which is bound to cause confusion. This overrides that definition in order to give a more helpful label to the port.

The Arduino IDE's override system gives precedence to the hardware package which has the vendor name that comes first in lexicographical order. Since the vendor name arduino comes before the vendor name MCUdude, I changed the vendor name of the microUPDIcore package to aaaMCUdude so that the definition in the microUPDIcore would override the definition in Arduino megaAVR Boards.